pyproject.toml
[project]
name = "self-evolving-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "weave>=0.51.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "crewai>=0.80.0",
    "crewai-tools>=0.15.0",
    "datasets>=2.0.0",
    "google-adk>=1.16.0",
    "daytona-sdk>=0.110.1",
    "pyyaml>=6.0.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
